requirements.txt
opencv-python
mediapipe
numpy
pyttsx3
torch
pyyaml
pandas
wandb